More Cygwin fixes.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@3339 42547f41-83fc-0310-9aef-9668e019ab9e

+++ util/mxput.c
@@ -114,7 +114,7 @@ main( int argc, char *argv[] )
 		} else {
 			if ( argv[i][0] == '-' ) {
 				if ( (argv[i][1] == '.')
-				  || isdigit( argv[i][1] ) )
+				  || isdigit( (int) (argv[i][1]) ) )
 				{
 					fixed_argv[i] = strdup( "--" );
